Addressed PR Concerns
